Fixed Protected Posts
If you use Next or Previous to navigate into a locked post the content 
is now hidden and the password form is shown. The meta information is 
still shown: dates, author(s), tags, and categories. Should this be 
hidden to? NOTE: Article comments are not working at all. They were 
removed a while ago and never put back in. This should be the next major 
project.