Alpha Release v0.0.1
This is the first truly usable in production version of Wiki Modern. 
Comments have not been implemented at all and will be missing. All other 
features expected to be in a theme are present. The code base needs a 
lot of refactoring and the Customizer works but is still buggy.